requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-openai
python-docx
python-dotenv
